requirements.txt
aiohttp==3.6.2
amqp==2.5.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
cachetools==4.1.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
email-validator==1.1.0
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
jikanpy==3.4.2
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
multidict==4.7.5
passlib==1.7.2
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
restless==2.1.1
rsa==4.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.5
uritemplate==3.0.1
urllib3==1.25.9
validators==0.15.0
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
yarl==1.4.2
